+++ src/ui/panes/mod.rs
@@ -2,12 +2,11 @@ use std::borrow::Cow;
 
 use album_art::AlbumArtPane;
 use albums::AlbumsPane;
-use anyhow::{Context, Result};
+use anyhow::Result;
 use artists::{ArtistsPane, ArtistsPaneMode};
-use crossterm::event::{KeyCode, KeyEvent};
+use crossterm::event::KeyEvent;
 use directories::DirectoriesPane;
 use either::Either;
-use itertools::Itertools;
 #[cfg(debug_assertions)]
 use logs::LogsPane;
 use playlists::PlaylistsPane;
@@ -16,19 +15,16 @@ use ratatui::{
     prelude::Rect,
     style::Style,
     text::{Line, Span},
-    widgets::ListItem,
     Frame,
 };
 use search::SearchPane;
 use strum::Display;
 
 use crate::{
-    cli::{create_env, run_external},
     config::{
-        keys::{CommonAction, GlobalAction},
+        keys::CommonAction,
         tabs::PaneType,
         theme::properties::{Property, PropertyKind, PropertyKindOrText, SongProperty, StatusProperty, WidgetProperty},
-        Config,
     },
     context::AppContext,
     mpd::{
@@ -38,11 +34,7 @@ use crate::{
     utils::{mouse_event::MouseEvent, DurationExt},
 };
 
-use super::{
-    utils::dirstack::{DirStack, DirStackItem},
-    widgets::volume::Volume,
-    KeyHandleResultInternal, UiEvent,
-};
+use super::{widgets::volume::Volume, KeyHandleResultInternal, UiEvent};
 
 pub mod album_art;
 pub mod albums;
@@ -88,12 +80,12 @@ impl PaneContainer {
             queue: QueuePane::new(context),
             #[cfg(debug_assertions)]
             logs: LogsPane::default(),
-            directories: DirectoriesPane::default(),
-            albums: AlbumsPane::default(),
-            artists: ArtistsPane::new(ArtistsPaneMode::Artist),
-            album_artists: ArtistsPane::new(ArtistsPaneMode::AlbumArtist),
-            playlists: PlaylistsPane::default(),
-            search: SearchPane::new(context.config),
+            directories: DirectoriesPane::new(context),
+            albums: AlbumsPane::new(context),
+            artists: ArtistsPane::new(ArtistsPaneMode::Artist, context),
+            album_artists: ArtistsPane::new(ArtistsPaneMode::AlbumArtist, context),
+            playlists: PlaylistsPane::new(context),
+            search: SearchPane::new(context),
             album_art: AlbumArtPane::new(context),
         }
     }
@@ -719,248 +711,6 @@ impl StringExt for String {
             ))
         } else {
             Cow::Borrowed(self)
-        }
-    }
-}
-
-enum MoveDirection {
-    Up,
-    Down,
-}
-
-#[allow(unused)]
-trait BrowserPane<T: DirStackItem + std::fmt::Debug>: Pane {
-    fn stack(&self) -> &DirStack<T>;
-    fn stack_mut(&mut self) -> &mut DirStack<T>;
-    fn set_filter_input_mode_active(&mut self, active: bool);
-    fn is_filter_input_mode_active(&self) -> bool;
-    fn next(&mut self, client: &mut impl MpdClient) -> Result<KeyHandleResultInternal>;
-    fn list_songs_in_item(&self, client: &mut impl MpdClient, item: &T) -> Result<Vec<Song>>;
-    fn move_selected(
-        &mut self,
-        direction: MoveDirection,
-        client: &mut impl MpdClient,
-    ) -> Result<KeyHandleResultInternal> {
-        Ok(KeyHandleResultInternal::SkipRender)
-    }
-    fn prepare_preview(
-        &mut self,
-        client: &mut impl MpdClient,
-        config: &Config,
-    ) -> Result<Option<Vec<ListItem<'static>>>>;
-    fn add(&self, item: &T, client: &mut impl MpdClient) -> Result<KeyHandleResultInternal>;
-    fn add_all(&self, client: &mut impl MpdClient) -> Result<KeyHandleResultInternal>;
-    fn delete(&self, item: &T, index: usize, client: &mut impl MpdClient) -> Result<KeyHandleResultInternal> {
-        Ok(KeyHandleResultInternal::SkipRender)
-    }
-    fn rename(&self, item: &T, client: &mut impl MpdClient) -> Result<KeyHandleResultInternal> {
-        Ok(KeyHandleResultInternal::SkipRender)
-    }
-    fn handle_filter_input(
-        &mut self,
-        event: KeyEvent,
-        client: &mut impl MpdClient,
-        config: &Config,
-    ) -> Result<KeyHandleResultInternal> {
-        match config.keybinds.navigation.get(&event.into()) {
-            Some(CommonAction::Close) => {
-                self.set_filter_input_mode_active(false);
-                self.stack_mut().current_mut().set_filter(None, config);
-                let preview = self.prepare_preview(client, config)?;
-                self.stack_mut().set_preview(preview);
-                Ok(KeyHandleResultInternal::RenderRequested)
-            }
-            Some(CommonAction::Confirm) => {
-                self.set_filter_input_mode_active(false);
-                self.stack_mut().current_mut().jump_next_matching(config);
-                let preview = self.prepare_preview(client, config)?;
-                self.stack_mut().set_preview(preview);
-                Ok(KeyHandleResultInternal::RenderRequested)
-            }
-            _ => match event.code {
-                KeyCode::Char(c) => {
-                    self.stack_mut().current_mut().push_filter(c, config);
-                    Ok(KeyHandleResultInternal::RenderRequested)
-                }
-                KeyCode::Backspace => {
-                    self.stack_mut().current_mut().pop_filter(config);
-                    Ok(KeyHandleResultInternal::RenderRequested)
-                }
-                _ => Ok(KeyHandleResultInternal::SkipRender),
-            },
-        }
-    }
-
-    fn handle_global_action(
-        &mut self,
-        action: GlobalAction,
-        client: &mut impl MpdClient,
-        context: &AppContext,
-    ) -> Result<KeyHandleResultInternal> {
-        match action {
-            GlobalAction::ExternalCommand { command, .. } if !self.stack().current().marked().is_empty() => {
-                let songs: Vec<_> = self
-                    .stack()
-                    .current()
-                    .marked_items()
-                    .map(|item| self.list_songs_in_item(client, item))
-                    .flatten_ok()
-                    .try_collect()?;
-                let songs = songs.iter().map(|song| song.file.as_str()).collect_vec();
-
-                run_external(command, create_env(context, songs, client)?);
-
-                Ok(KeyHandleResultInternal::SkipRender)
-            }
-            GlobalAction::ExternalCommand { command, .. } => {
-                if let Some(selected) = self.stack().current().selected() {
-                    let songs = self.list_songs_in_item(client, selected)?;
-                    let songs = songs.iter().map(|s| s.file.as_str());
-
-                    run_external(command, create_env(context, songs, client)?);
-                }
-                Ok(KeyHandleResultInternal::SkipRender)
-            }
-            _ => Ok(KeyHandleResultInternal::KeyNotHandled),
-        }
-    }
-
-    fn handle_common_action(
-        &mut self,
-        action: CommonAction,
-        client: &mut impl MpdClient,
-        context: &AppContext,
-    ) -> Result<KeyHandleResultInternal> {
-        let config = context.config;
-        match action {
-            CommonAction::Up => {
-                self.stack_mut().current_mut().prev();
-                let preview = self.prepare_preview(client, config).context("Cannot prepare preview")?;
-                self.stack_mut().set_preview(preview);
-                Ok(KeyHandleResultInternal::RenderRequested)
-            }
-            CommonAction::Down => {
-                self.stack_mut().current_mut().next();
-                let preview = self.prepare_preview(client, config).context("Cannot prepare preview")?;
-                self.stack_mut().set_preview(preview);
-                Ok(KeyHandleResultInternal::RenderRequested)
-            }
-            CommonAction::MoveUp => {
-                let res = self.move_selected(MoveDirection::Up, client)?;
-                Ok(res)
-            }
-            CommonAction::MoveDown => {
-                let res = self.move_selected(MoveDirection::Down, client)?;
-                Ok(res)
-            }
-            CommonAction::DownHalf => {
-                self.stack_mut().current_mut().next_half_viewport();
-                let preview = self.prepare_preview(client, config).context("Cannot prepare preview")?;
-                self.stack_mut().set_preview(preview);
-                Ok(KeyHandleResultInternal::RenderRequested)
-            }
-            CommonAction::UpHalf => {
-                self.stack_mut().current_mut().prev_half_viewport();
-                let preview = self.prepare_preview(client, config).context("Cannot prepare preview")?;
-                self.stack_mut().set_preview(preview);
-                Ok(KeyHandleResultInternal::RenderRequested)
-            }
-            CommonAction::Bottom => {
-                self.stack_mut().current_mut().last();
-                let preview = self.prepare_preview(client, config).context("Cannot prepare preview")?;
-                self.stack_mut().set_preview(preview);
-                Ok(KeyHandleResultInternal::RenderRequested)
-            }
-            CommonAction::Top => {
-                self.stack_mut().current_mut().first();
-                let preview = self.prepare_preview(client, config).context("Cannot prepare preview")?;
-                self.stack_mut().set_preview(preview);
-                Ok(KeyHandleResultInternal::RenderRequested)
-            }
-            CommonAction::Right => {
-                let res = self.next(client)?;
-                let preview = self.prepare_preview(client, config).context("Cannot prepare preview")?;
-                self.stack_mut().set_preview(preview);
-                Ok(res)
-            }
-            CommonAction::Left => {
-                self.stack_mut().pop();
-                let preview = self.prepare_preview(client, config).context("Cannot prepare preview")?;
-                self.stack_mut().set_preview(preview);
-                Ok(KeyHandleResultInternal::RenderRequested)
-            }
-            CommonAction::EnterSearch => {
-                self.set_filter_input_mode_active(true);
-                self.stack_mut().current_mut().set_filter(Some(String::new()), config);
-                Ok(KeyHandleResultInternal::RenderRequested)
-            }
-            CommonAction::NextResult => {
-                self.stack_mut().current_mut().jump_next_matching(config);
-                let preview = self.prepare_preview(client, config).context("Cannot prepare preview")?;
-                self.stack_mut().set_preview(preview);
-                Ok(KeyHandleResultInternal::RenderRequested)
-            }
-            CommonAction::PreviousResult => {
-                self.stack_mut().current_mut().jump_previous_matching(config);
-                let preview = self.prepare_preview(client, config).context("Cannot prepare preview")?;
-                self.stack_mut().set_preview(preview);
-                Ok(KeyHandleResultInternal::RenderRequested)
-            }
-            CommonAction::Select => {
-                self.stack_mut().current_mut().toggle_mark_selected();
-                self.stack_mut().current_mut().next();
-                let preview = self.prepare_preview(client, config).context("Cannot prepare preview")?;
-                self.stack_mut().set_preview(preview);
-                Ok(KeyHandleResultInternal::RenderRequested)
-            }
-            CommonAction::Add if !self.stack().current().marked().is_empty() => {
-                for idx in self.stack().current().marked().iter().rev() {
-                    let item = &self.stack().current().items[*idx];
-                    self.add(item, client)?;
-                }
-                Ok(KeyHandleResultInternal::RenderRequested)
-            }
-            CommonAction::Add => {
-                if let Some(item) = self.stack().current().selected() {
-                    self.add(item, client)
-                } else {
-                    Ok(KeyHandleResultInternal::SkipRender)
-                }
-            }
-            CommonAction::AddAll if !self.stack().current().items.is_empty() => {
-                self.add_all(client)?;
-                Ok(KeyHandleResultInternal::RenderRequested)
-            }
-            CommonAction::AddAll => Ok(KeyHandleResultInternal::SkipRender),
-            CommonAction::Delete if !self.stack().current().marked().is_empty() => {
-                for idx in self.stack().current().marked().iter().rev() {
-                    let item = &self.stack().current().items[*idx];
-                    self.delete(item, *idx, client)?;
-                }
-                Ok(KeyHandleResultInternal::RenderRequested)
-            }
-            CommonAction::Delete => {
-                if let Some((index, item)) = self.stack().current().selected_with_idx() {
-                    self.delete(item, index, client)?;
-                    Ok(KeyHandleResultInternal::RenderRequested)
-                } else {
-                    Ok(KeyHandleResultInternal::SkipRender)
-                }
-            }
-            CommonAction::Rename => {
-                if let Some(item) = self.stack().current().selected() {
-                    self.rename(item, client)
-                } else {
-                    Ok(KeyHandleResultInternal::SkipRender)
-                }
-            }
-            CommonAction::FocusInput => Ok(KeyHandleResultInternal::SkipRender),
-            CommonAction::Close => Ok(KeyHandleResultInternal::SkipRender), // todo out?
-            CommonAction::Confirm => Ok(KeyHandleResultInternal::SkipRender), // todo next?
-            CommonAction::PaneDown => Ok(KeyHandleResultInternal::SkipRender),
-            CommonAction::PaneUp => Ok(KeyHandleResultInternal::SkipRender),
-            CommonAction::PaneRight => Ok(KeyHandleResultInternal::SkipRender),
-            CommonAction::PaneLeft => Ok(KeyHandleResultInternal::SkipRender),
         }
     }
 }