+++ src/ui/panes/playlists.rs
@@ -1,7 +1,11 @@
 use anyhow::{anyhow, Context, Result};
 use crossterm::event::KeyEvent;
 use itertools::Itertools;
-use ratatui::{prelude::Rect, widgets::ListItem, Frame};
+use ratatui::{
+    prelude::Rect,
+    widgets::{ListItem, StatefulWidget},
+    Frame,
+};
 
 use crate::{
     config::Config,
@@ -11,34 +15,43 @@ use crate::{
         mpd_client::{Filter, MpdClient, SingleOrRange, Tag},
     },
     ui::{
+        browser::{BrowserPane, MoveDirection},
         modals::rename_playlist::RenamePlaylistModal,
         utils::dirstack::{DirStack, DirStackItem},
         widgets::browser::Browser,
         KeyHandleResultInternal, UiEvent,
     },
-    utils::macros::{status_error, status_info},
+    utils::{
+        macros::{status_error, status_info},
+        mouse_event::MouseEvent,
+    },
 };
 
-use super::{browser::DirOrSong, BrowserPane, Pane};
+use super::{browser::DirOrSong, Pane};
 
 #[cfg(test)]
 mod tests;
 
-#[derive(Debug, Default)]
+#[derive(Debug)]
 pub struct PlaylistsPane {
     stack: DirStack<DirOrSong>,
     filter_input_mode: bool,
+    browser: Browser<DirOrSong>,
+}
+
+impl PlaylistsPane {
+    pub fn new(context: &AppContext) -> Self {
+        Self {
+            stack: DirStack::default(),
+            filter_input_mode: false,
+            browser: Browser::new(context.config),
+        }
+    }
 }
 
 impl Pane for PlaylistsPane {
-    fn render(&mut self, frame: &mut Frame, area: Rect, AppContext { config, .. }: &AppContext) -> Result<()> {
-        frame.render_stateful_widget(
-            Browser::new(config)
-                .set_widths(&config.theme.column_widths)
-                .set_border_style(config.as_border_style()),
-            area,
-            &mut self.stack,
-        );
+    fn render(&mut self, frame: &mut Frame, area: Rect, _context: &AppContext) -> Result<()> {
+        self.browser.render(area, frame.buffer_mut(), &mut self.stack);
 
         Ok(())
     }
@@ -131,6 +144,15 @@ impl Pane for PlaylistsPane {
             }
             _ => Ok(KeyHandleResultInternal::SkipRender),
         }
+    }
+
+    fn handle_mouse_event(
+        &mut self,
+        event: MouseEvent,
+        client: &mut impl MpdClient,
+        context: &mut AppContext,
+    ) -> Result<KeyHandleResultInternal> {
+        self.handle_mouse_action(event, client, context)
     }
 
     fn handle_action(
@@ -260,7 +282,7 @@ impl BrowserPane<DirOrSong> for PlaylistsPane {
 
     fn move_selected(
         &mut self,
-        direction: super::MoveDirection,
+        direction: MoveDirection,
         client: &mut impl MpdClient,
     ) -> Result<KeyHandleResultInternal> {
         let Some((idx, selected)) = self.stack().current().selected_with_idx() else {
@@ -275,8 +297,8 @@ impl BrowserPane<DirOrSong> for PlaylistsPane {
             DirOrSong::Dir { .. } => {}
             DirOrSong::Song(_) => {
                 let new_idx = match direction {
-                    super::MoveDirection::Up => idx.saturating_sub(1),
-                    super::MoveDirection::Down => (idx + 1).min(self.stack().current().items.len() - 1),
+                    MoveDirection::Up => idx.saturating_sub(1),
+                    MoveDirection::Down => (idx + 1).min(self.stack().current().items.len() - 1),
                 };
                 client.move_in_playlist(playlist, &SingleOrRange::single(idx), new_idx)?;
             }
@@ -307,5 +329,9 @@ impl BrowserPane<DirOrSong> for PlaylistsPane {
                         .collect_vec(),
                 }))
             })
+    }
+
+    fn browser_areas(&self) -> [Rect; 3] {
+        self.browser.areas
     }
 }